WL#15135 post-push fix
Add stub functions to allow NodeCertificate.cpp to compile
succesfully on EL7 with OpenSSL 1.0.2.

+++ storage/ndb/src/common/util/testTlsKeyManager.cpp
@@ -44,7 +44,7 @@
 #include "util/TlsKeyManager.hpp"
 
 int opt_port = 4400;
-short opt_last_test = INT16_MAX;
+int opt_last_test = INT16_MAX;
 
 static struct my_option options[] =
 {
